Fixed compilation on MacOS platforms where Python default target is <10.9.
Modified setup.py to set env MACOSX_DEPLOYMENT_TARGET=10.9 when the build
MacOS is that or newer, but the default MACOSX_DEPLOYMENT_TARGET (when absent)
is <10.9, which happens when the running Python targets <10.9 min.

This happens for example when running a current Anaconda Python, which
has target 10.7.

This fix taken verbatim from similar in Pandas,
https://github.com/pandas-dev/pandas/pull/24274 .

Fixed #39.