Mark uses of ragged nested sequences in NumPy arrays as intented for two tests
The tests check that emd() raises a ValueError when passed incorrect matrices.

The ragged sequences are now deprecated in NumPy unless marked as intended:

VisibleDeprecationWarning: Creating an ndarray from ragged nested sequences
(which is a list-or-tuple of lists-or-tuples-or ndarrays with different
lengths or shapes) is deprecated. If you meant to do this, you must
specify 'dtype=object' when creating the ndarray.